Move and rename sample data file
